Switch to float64

+++ internal/models/http.go
@@ -5,7 +5,7 @@ type ListProduct struct {
 	Name          string  `json:"name"`
 	Description   string  `json:"description"`
 	ThumbnailURL  string  `json:"thumbnail_url"`
-	AverageRating float32 `json:"average_rating"`
+	AverageRating float64 `json:"average_rating"`
 }
 
 type Products struct {
