+++ internal/transport/http/incoming.go
@@ -3,5 +3,5 @@ package http
 type AddProductReviewRequest struct {
 	Name   string  `validate:"required"   json:"name"`
 	Text   string  `                      json:"text"`
-	Rating float32 `validate:"gt=0,lte=5" json:"rating"`
+	Rating float64 `validate:"gt=0,lte=5" json:"rating"`
 }
